Implemented encryption algorithm without block cipher
pad10 implements the padding function as described in the paper.

XOR_block is a blockwise xor function. It behaves like a pipeline where the output gets computed and the inputs aren't available afterwards. This is closer to a possible hardware implementation at the expense of software efficiency.

The Algorithm

The actual algorithm has 5 stages: The first is computing the partial tag from the additional data, the second same but from the secret message, the third is the actual computation of the tag from the two stages before, the fourth encrypts the message and the fifth combines ciphertext and tag into one packet.

The first and second stage do the same computation, but they work with different data and slightly different constants. In the function compute_partial_tag the data blocks are xored with a cipher stream using a tweak based on the block number. In the event that the data doesn't evenly divide into blocks, the last (partial) data block is padded using the beforementioned pad10 function.

The third stage combines the two partial tags and encrypts it using the nonce as a tweak. Since the nonce is only 120 bits long, the first 8 bit are constant.

The fourth stage is where the encryption of the secret message happens. The message is encrypted using the stream cipher scheme (like CTR in AES). To make later computations easier, partial data blocks are padded with zeroes. The length must be saved in some other way. The padding is done by copying the message into a multiple-of-block-size buffer of zeroes.

The stream cipher is obtained by encrypting the nonce with different tweaks based on the block number, similar to the CTR mode in AES. However, the tweak also depends on the computed tag, making multiple usages of the same nonce less severe.

In the final stage, the computed ciphertext and the tag get concatenated to form a single unit.

The authenticated ciphertext can now be send to a decryptor.

Also defined constants relating to the bytelength of keys, nonces, tags and tweaks in the header. Added safeguard against multiple declaration.