Updated README + Correction of last commit message
Added a 'Goals' section to describe the goals and the limitations of the project.

Added a 'Disclaimers' section to describe limitations of development and explicit issues that must be expected when using the code.

Correction of last commit: In the first and second stage, the data blocks are not xored like a stream cipher, but are themselves encrypted using a counter-style tweak. The encrypted blocks are then xored together to produce the tag. Should a block not evenly fit, it must be padded to be encrypted using the block cipher. With a stream cipher, this would not be necessary.