Implement block cipher Deoxys-BC-384
The block cipher works similar to AES.

The 384-bit version uses 16 rounds that work almost identical to AES, only the round keys are computed differently. The AES-like permutations and substitutions are taken from a pre-existing AES implementation by kokke.

The internal state, initialized with the plaintext, is XORed with the round keys (called subtweakeys) and a round constant. There are 3 subtweakeys per round which depend on the key and tweak. The subtweakeys are changed after each round by shifting them and permuting them using a permutation called h.