Use block cipher in Deoxys, fix minor issues
I found some test cases for Deoxys. They are declared as official, but I cannot check that currently

XOR function now takes additional parameter length instead of hardcoded BLOCK_LEN. In practice, the inputs are always of length BLOCK_LEN (because TAG_LEN is the same length).

In C, to do bitwise OR, a single ¦ is enough, the double ¦¦ is a boolean operator, which is wrong.