Compute affine transformation in SubBytes
An implementation of the matrix multiplication plus constant used for the substitution step.

The inversion of the input is still done via a LUT.

The pure LUT-version is about 20% faster, which is probably due to the ton of shifts and operations that need to be done.

In hardware, shifts, bit selection and xor are trivial operations, while LUT are relatively expensive, so this might be interesting for a future hardware implementation.