Implement inversion algorithm for Sbox calculation
Previously, the Sbox would be calculated using a LUT for multiplicative inverses. This is now replaced with a computational solution.

The algorithm is explained with further detail in the document 'fast_hardware_inversion.pdf'. As the name suggests, it is intended for hardware implementation and is one of the fastest in that regard.

In software implemented though, the algorithm takes almost 3x as long due to the many logical operations and subsequent array accesses.

Also, there are many design choices made regarding readability in order to make comparisons with the source material easier (e.g. bit arrays padded with 0 to start at index 1)

The algorithm itself is done in several stages. Firstly, there is a conversion into an isomorphic field GF(256) with a different polynomial. Second is the split into two arrays containing the high and low bits respectively. Third are the 3 actual stages of computation, where the inverse is computed using smaller subfields and different representations. And lastly, the result is converted back into the field of AES, while simultaneously computing the matrix multiplication. In the end, the result is XORed with the constant 0x63 to account for the full affine transformation.