Use stack-only allocation in DeoxysBC implementation
The previous implementation used heap allocation in order to keep to a more functional style of programming that is also similar to a hardware implementation

This turned out to be expensive. The version without vector extensions sees a performance improvement of 50% since it heavily relied on that.

This refactoring not only improves readability by restricting the scope of the allocations, it also allowed the removal of some assignments because the values are modified in place (more imperative style of programming). This change lead to a performance improvement of about 10% that is enjoyed by both versions.