Add function Deoxys_BC_encrypt_buffer with buffer for result
The new function performs the same operation as the function of similar name, but writes the result into the given buffer instead of creating a new one.

The other function is implemented on top of the new one, needing calloc for that.

Also refactored the code using vector extensions. The new function has to be implemented with all 4 combinations of USE_AES_NI and GCC_VECTOR_EXTENSIONS. At this point it might be better to simply use different files.