Reuse tweakeys depending on the key
The tweakeys tk2 and tk3 solely depend on the key and their key schedules. As such, they can be precomputed and then read during execution.

This not only improves performance by up to 30%, but also simplifies the round function with a simple XOR instead of the complex addsubtweakey function.

The previous function Deoxys_BC_encrypt_buffer has been implemented on top of the new function, generating the subtweakeys on the fly without storing them.

The idea for this optimization came from the Deoxys paper. This optimization is a significant break from the close-to-hardware model, but is more suited for the software implementation.